Adjust Cynthia's Grandma Coords
Moved the HM03 from Cynthia's Grandmother to be on the cave entrance to the Celestic Ruins.  This run, the story flags must have aligned, because I was able to get the event to trigger.  Might have to do with clearing certain Galactic things?  Or gym badges?  Not sure, but for sure there is a plot flag somewhere in the code that needs to be triggered for this event to actually occur.